clear remaining codacy and sonar findings
