+++ scripts/provider_ui/provider_admin_bootstrap.mjs
@@ -352,5 +352,9 @@ function reportCliError(error) {
 }
 
 if (isCliEntrypoint(import.meta.url)) {
-  main().catch(reportCliError);
+  try {
+    await main();
+  } catch (error) {
+    reportCliError(error);
+  }
 }
